refactor(project): merge project_create.go into project.go
Single file per resource for now — will split if it grows.
No behavior change; tests still pass; live smoke-test against
skipi.semaphoreci.com on ~/Workspace/calculator still returns
status:"exists" for the existing project.
